WIP: Working call expressions
This adds separate expression types in the AST for both method calls and
property access on a struct. Currently method calls do not work with
arguments, however the machinery is in place to support it.

Other outstanding issues are structs with methods using pointer
receivers also seem not to work, despite attempts to use
reflect.Indirect. Further investigation is needed.